message-to-command/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
appdirs==1.4.3            # via setuptools
chainer==1.22.0
cycler==0.10.0            # via matplotlib
filelock==2.0.8           # via chainer
jaconv==0.2.0
matplotlib==2.0.0
mecab-python3==0.7
nltk==3.2.2
nose==1.3.7               # via chainer
numpy==1.12.1
olefile==0.44             # via pillow
packaging==16.8           # via setuptools
pillow==4.0.0
protobuf==3.2.0           # via chainer
pyparsing==2.2.0          # via matplotlib, packaging
python-dateutil==2.6.0    # via matplotlib
pytz==2017.2              # via matplotlib
scikit-learn==0.18.1
scipy==0.19.0
six==1.10.0               # via chainer, cycler, matplotlib, nltk, packaging, protobuf, python-dateutil, setuptools

# The following packages are considered to be unsafe in a requirements file:
# setuptools                # via protobuf
